Got pytest to work on GPUs
