+++ pytest-benchmark.ini
@@ -1,0 +1,5 @@
+# pytest-benchmark.ini
+[pytest]
+minversion = 6.0
+addopts = -v --durations=0 --benchmark-autosave
+python_files = htf/test-py/benchmark.py
